Added some more validation and tests to value modification controller

+++ src/controllers/UserAwareController.java
@@ -9,8 +9,8 @@ public abstract class UserAwareController extends Controller {
 
   private static final ImmutableSet<String> ADMIN_ROLES = ImmutableSet.of("admin", "anonymous");
   private static final ImmutableSet<String> ANONYMOUS_ROLES = ImmutableSet.of("anonymous");
-  public static final Short MINIMUM_YEAR = 2014;
-  public static final Short MAXIMUM_YEAR = 2020;
+  public static final Integer MINIMUM_YEAR = 2014;
+  public static final Integer MAXIMUM_YEAR = 2020;
 
   public String getLoggedInUsername() {
     return (String) session.getAttribute("username");
